Делаю

+++ Assets/InternalAssets/Characters/Base/Stats/AttackStatsConfig.cs
@@ -9,6 +9,6 @@ namespace SouthBasement.Weapons
         [field: SerializeField] public int Damage { get; private set; } = 12;
         [field: SerializeField] public float AttackRange { get; private set; } = 0.4f;
         [field: SerializeField] public float AttackRate { get; private set; } = 1f;
-        [field: SerializeField] public int StaminaRequire { get; private set; } = 10;
+        [field: SerializeField] public float StaminaRequire { get; private set; } = 10;
     }
 }
